feat: SearchKeyControl
Added a new component: SearchKeyControl, which leverages qIndex to allow specific toggling on/off of searchKeys.

Also added a dynamic rendering utility function, which allows the rendering of items in dynamic rows. In the case of SearchKey control it will allow us to add/remove key options dynamically and the checkbox rows will scale in kind.

ERM-2290