fix: hasInterface
Fixed issue where hasInterface returns an unexpected value (hasInterface returns the interface value in the system if the condition is met, or `0` otherwise. This forces us us to use the operator `!!` to get a boolean value out of it.

In addition, ERM has made a point of supporting both users interface 15 AND interface 16, and the hasInterface block has been expanded to reflect this.

ERM-2390