ERM-2533: Refactor interfaces code to react-query (#577)
* feat: Interface hooks (!WIP!)

Started work on moving interface requirements over to react-query.
This involves multiple hooks and rewrangling parts of the display to fetch by themselves.

THIS IS A WIP COMMIT AND DOES NOT CURRENTLY FUNCTION. DO NOT MERGE AS IS

refs ERM-2533

* feat: useSingleFetchInterfaceCredentials

Fixed interfaces, exposed new hook which simplifies the "fetch one at a time and return an object with all of the credential data in place" logic

Also some light refactors of ViewOrganizationCard

ERM-2533

* test: Fixed tests

Fixed tests for ViewOrganizationCard

ERM-2533

* chore: SonarLint