feat: SearchKeyControl
Some futureproofing integration between usePrevNextPagination and useQindex, to ensure that a change in qIndex resets the page.

To that end, add a prop defaultToPageOne to usePrevNextPagination which (when set to false) allows the reading of currentPage without setting the param if undefined, effectively being able to read if page param is in use or not without affecting it.